+++ commons-server/src/main/kotlin/de/gematik/zeta/zetaguard/keycloak/commons/smcb/ZetaGuardTokenExchangeData.kt
@@ -1,0 +1,43 @@
+/*-
+ * #%L
+ * keycloak-zeta
+ * %%
+ * (C) akquinet tech@Spree GmbH, 2025, licensed for gematik GmbH
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * *******
+ *
+ * For additional notes and disclaimer from gematik and in case of changes by gematik find details in the "Readme" file.
+ * #L%
+ */
+package de.gematik.zeta.zetaguard.keycloak.commons.smcb
+
+import java.beans.ConstructorProperties
+import java.time.Duration
+
+/**
+ * Data gathered during token exchange and OPA decision process.
+ *
+ * Used, e.g., to generate access and refresh tokens
+ */
+data class ZetaGuardTokenExchangeData
+@ConstructorProperties("telematikID", "professionOID", "subjectOrganisation", "subjectName", "accessTokenTTL", "refreshTokenTTL")
+constructor(
+    val telematikID: String,
+    val professionOID: String,
+    val subjectOrganisation: String,
+    val subjectName: String,
+    val accessTokenTTL: Duration,
+    val refreshTokenTTL: Duration,
+)